ZA: Move the interests related classes in views.py to their own file
